GCP DEPLOYED CROP DISEASE CLASSIFICATION REST API
This folder is deployed on gcp as a REST API and the url is used to make a request from the android application to classify the crop disease but sending the image and getting output using POST request. 

Note: POST request is used instead of GET request here because the data sent to the cloud is an image file and when sending large image files, the process was slower and lossy.